+++ src_3_0_0/modules/routing/js/directions/routing-ors-directions.js
@@ -73,10 +73,10 @@ async function fetchRoutingOrsDirections ({
         });
     }
     catch (e) {
-        if (e.response.status === 404) {
+        if (e.response?.status === 404) {
             throw new Error(i18next.t("common:modules.routing.errors.noRouteFound"));
         }
-        if (e.response && e.response.data && e.response.data.error) {
+        if (e.response?.data?.error) {
             if (e.response.data.error.code === 2003) {
                 throw new Error(i18next.t("common:modules.routing.errors.avoidAreaBig"));
             }